requirements.txt
Django>=4.1.3
gunicorn
requests
django-dotenv
psycopg2-binary
django-storages
boto3